+++ test/index.js
@@ -1,0 +1,6 @@
+console.warn('[test/index.js]', new Date().toISOString(), '\n----------------------------------------------------------\n');
+
+require('babel-register');
+
+require('./callFactory.test');
+require('./createStore.test');